Add browserify 3.x support
`util.isBuffer(new Buffer(4))` currently returns false with
`core-util-is` and browserify 3.x.

In short, `instanceof Buffer` doesn’t work to check if an object is a
`Buffer` because in browserify, the Buffer constructor actually returns
an instance of `Uint8Array` for performance reasons.

So, by checking with `Buffer.isBuffer` instead, that gives browserify a
chance to define it’s own Buffer.isBuffer implementation. I believe
that in node, `Buffer.isBuffer` just does `instanceof` anyway.